requirements.txt
python-telegram-bot==21.3
anthropic>=0.40.0
python-dotenv>=1.0.0
playwright>=1.44.0
asyncpg>=0.29.0
aiohttp>=3.9.0
httpx>=0.27.0
